requirements.txt
# LangChain Core
langchain
langchain-core
langchain_community

# Models & Embeddings
langchain-groq
google-generativeai
fastembed

# Environment Variable Management
python-dotenv

# Document Loaders & Processors
pypdf
pymupdf
langdetect
docx2txt
rapidocr-onnxruntime
requests

# Frontend
reflex==0.8.27
gunicorn
reportlab
chromadb